% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ont_annot.R
\name{.can_use_network}
\alias{.can_use_network}
\title{Internal - should we allow network access now?}
\usage{
.can_use_network()
}
\value{
Logical scalar.
}
\description{
Controlled by the R option \code{tidyexposomics.allow_network} and \code{interactive()}.
}
\keyword{internal}
