% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ont_annot.R
\name{.get_ols_description}
\alias{.get_ols_description}
\title{Internal - fetch OLS description for a term}
\usage{
.get_ols_description(ontology_id, ontology_prefix)
}
\arguments{
\item{ontology_id}{Term ID (e.g., \code{"HP:0004322"}).}

\item{ontology_prefix}{Prefix (e.g., \code{"HP"}).}
}
\value{
A character scalar description or \code{NA_character_}.
}
\description{
Internal - fetch OLS description for a term
}
\keyword{internal}
