% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ont_annot.R
\name{.load_ontologies}
\alias{.load_ontologies}
\title{Internal - load ontologies}
\usage{
.load_ontologies(use_demo = TRUE, ...)
}
\arguments{
\item{use_demo}{Logical; if TRUE, load packaged demo objects \code{hpo}, \code{ecto}, \code{chebi}.}

\item{...}{Optional named overrides: \code{hpo}, \code{ecto}, \code{chebi}.}
}
\value{
A list with elements \code{hpo}, \code{ecto}, \code{chebi}.
}
\description{
Internal - load ontologies
}
\keyword{internal}
