% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Internal - onLoad hook to register www assets}
\usage{
.onLoad(libname, pkgname)
}
\value{
Invisibly returns \code{NULL}
}
\description{
Registers \code{inst/app/www} as a Shiny resource path \code{"www"} if present.
}
\keyword{internal}
