% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_ont_annot.R
\name{.ont_annot_build_server}
\alias{.ont_annot_build_server}
\title{Internal - builds ontology annotation server}
\usage{
.ont_annot_build_server(ontologies)
}
\arguments{
\item{ontologies}{A list with \code{hpo}, \code{ecto}, \code{chebi} data.frames.}
}
\value{
A Shiny server function.
}
\description{
Internal - builds ontology annotation server
}
\keyword{internal}
