% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_ont_annot.R
\name{.ont_annot_build_ui}
\alias{.ont_annot_build_ui}
\title{Internal - builds ontology annotation UI}
\usage{
.ont_annot_build_ui()
}
\value{
A \code{shiny.tag} UI definition.
}
\description{
Internal - builds ontology annotation UI
}
\keyword{internal}
