% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{download_dataset}
\alias{download_dataset}
\title{Download and cache a tidyexposomics dataset}
\usage{
download_dataset(
  name = c("omics_list", "fdata", "meta", "annotated_cb", "expom_1", "chebi", "ecto",
    "hpo"),
  verbose = TRUE
)
}
\arguments{
\item{name}{Dataset name: one of "omics_list", "fdata", "meta",
"annotated_cb", "expom_1", "chebi", "ecto", "hpo".}

\item{verbose}{Logical; print messages.}
}
\value{
A list or object loaded from the cached .RData.
}
\description{
Download and cache a tidyexposomics dataset
}
