% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_annotation_data}
\alias{load_annotation_data}
\title{Load Ontology Data}
\usage{
load_annotation_data(
  to_load = c("all", "chebi", "ecto", "hpo"),
  verbose = TRUE,
  validate = TRUE
)
}
\arguments{
\item{to_load}{Character vector indicating which ontology to load.}

\item{verbose}{Logical; print messages.}

\item{validate}{Logical; validate MD5 checksum.}
}
\value{
A named list of ontology objects.
}
\description{
Downloads and loads \code{chebi}, \code{ecto}, and \code{hpo}.
}
\examples{
\dontrun{
# load single ontology
onts <- load_annotation_data(to_load = "ecto")

# load all ontologies
onts <- load_annotation_data(to_load = "all")
}

}
