% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ont_annot_app.R
\name{ont_annot_app}
\alias{ont_annot_app}
\title{Launch Ontology Annotation Shiny App}
\usage{
ont_annot_app()
}
\value{
Launches a Shiny app in a new window; does not return a value.
}
\description{
This function launches the Shiny application for annotating exposome variable names
using ontology terms such as HPO, ENVO, and ChEBI. The app is bundled with the
\code{tidyexposomics} package and allows interactive selection or search of ontology terms.
}
\examples{
if (interactive()) {
  ont_annot_app()
}
}
