% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_summary.R
\name{plot_missing_summary}
\alias{plot_missing_summary}
\title{Plot Summary of Missing Data Across Assays}
\usage{
plot_missing_summary(expomicset, threshold = 20)
}
\arguments{
\item{expomicset}{A \code{MultiAssayExperiment} object with missing data summary stored in \code{metadata(expomicset)$na_qc}.}

\item{threshold}{A numeric value specifying the missing data percentage threshold for features to be counted. Default is \code{20}.}
}
\value{
A \code{ggplot} object showing missing data statistics across assays.
}
\description{
Visualizes the number of features exceeding a missing data threshold in each assay of a \code{MultiAssayExperiment} object.
}
\details{
This function extracts missing data summaries from \code{metadata(expomicset)$na_qc}, generated by \code{filter_missing()},
and counts the number of features in each assay exceeding the specified missingness threshold. The results are:
\itemize{
\item Displayed as a \strong{bar plot}, where bars represent different assays.
\item Labeled with the number of features exceeding the threshold.
\item Colored by assay name for easy distinction.
}
}
\examples{
\dontrun{
plot_missing_summary(expom, threshold = 10)
}

}
