% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_clusters.R
\name{plot_sample_clusters}
\alias{plot_sample_clusters}
\title{Plot Sample Clusters}
\usage{
plot_sample_clusters(expomicset, cols_of_interest = NULL)
}
\arguments{
\item{expomicset}{A \code{MultiAssayExperiment} object containing sample clustering results
in \code{metadata(expomicset)$sample_clustering}.}

\item{cols_of_interest}{A character vector specifying columns from \code{colData} to include
in the summary. Default is \code{NULL}, which includes all available columns.}
}
\value{
A \code{ComplexHeatmap} plot displaying sample clustering results.
}
\description{
Generates a heatmap of sample clustering results and summarizes sample group assignments.
}
\details{
This function:
\itemize{
\item Extracts sample cluster assignments from \code{metadata(expomicset)$sample_clustering}.
\item Merges cluster labels with \code{colData(expomicset)}.
\item Plots the heatmap stored in \code{metadata(expomicset)$sample_clustering$heatmap}.
}
}
\examples{
\dontrun{
plot_sample_clusters(expom)
}

}
