% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_expomicset.R
\name{create_expomicset}
\alias{create_expomicset}
\title{Create an expomicset Object}
\usage{
create_expomicset(var_info, exposure, omics, row_data = NULL)
}
\arguments{
\item{var_info}{A data frame containing variable information metadata.}

\item{exposure}{A data frame containing exposure data, with rows as samples and columns as variables.}

\item{omics}{A list of matrices or a single matrix representing omics data. Each matrix should have samples as columns and features as rows.}

\item{row_data}{An optional list of \code{DataFrame} objects providing feature metadata for each omics dataset. If \code{NULL}, row metadata is generated automatically. Default is \code{NULL}.}
}
\value{
A \code{MultiAssayExperiment} object containing the formatted exposure and omics datasets.
}
\description{
Constructs a \code{MultiAssayExperiment} object from exposure data and omics datasets,
ensuring proper formatting and alignment of samples and features.
}
\details{
The function validates inputs, converts \code{omics} into a list if necessary, ensures all datasets are matrices with column names,
and creates \code{SummarizedExperiment} objects for each omics dataset. It then constructs a \code{MultiAssayExperiment} object
with exposure data in \code{colData} and variable information stored in metadata.
}
\examples{
\dontrun{
expomicset <- create_expomicset(
  var_info = des,
  exposure = meta,
  omics = list(mRNA = mrna_mat, proteomics = proteomics_mat)
)
}

}
