% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{download_dataset}
\alias{download_dataset}
\title{Download tidyexposomics zenodo data}
\usage{
download_dataset(
  name = c("omics_list", "fdata", "meta", "annotated_cb", "chebi", "ecto", "hpo"),
  dest_dir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{name}{The name of the dataset (without .zip or .RData extension).
Valid options include: "omics_list", "fdata", "meta", "annotated_cb",
"chebi", "ecto", "hpo".}

\item{dest_dir}{Destination directory to save
and extract the file (default: tempdir()).}

\item{verbose}{Print messages? (default: TRUE)}
}
\value{
Invisibly returns the loaded object.
}
\description{
Download and load a zipped .RData file from the tidyexposomics
Zenodo release: https://doi.org/10.5281/zenodo.17049349. The data bundle
contains ontology data from ECTO, CHEBI, and HPO, and processed data from
the ISGlobal Exposome Data Challenge 2021
(Maitre et al., Environment International, 2022;
DOI: 10.1016/j.envint.2022.107422). Please cite the original study
and data creators when using this bundle.
}
\examples{
# load the example meta data
download_dataset(
    name = "meta"
)

}
