% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{load_example_data}
\alias{load_example_data}
\title{Load example omics data for vignette/demo use}
\usage{
load_example_data(dest_dir = tempdir())
}
\arguments{
\item{dest_dir}{Destination directory to store temporary data files.}
}
\description{
Downloads and loads \code{omics_list}, \code{annotated_cb}, \code{fdata}, and \code{meta} into the global environment.
}
