% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_create_network.R
\name{run_create_network}
\alias{run_create_network}
\title{Create and store a network from correlation results in a MultiAssayExperiment}
\usage{
run_create_network(
  expomicset,
  feature_type = c("degs", "omics", "factors", "exposures"),
  action = c("add", "get")
)
}
\arguments{
\item{expomicset}{A \code{MultiAssayExperiment} object.}

\item{feature_type}{One of "degs", "omics", "factors", or "exposures". Used to select correlation results.}

\item{action}{"add" to store the graph in metadata, or "get" to return it.}
}
\value{
If \code{action = "add"}, returns modified \code{MultiAssayExperiment}. Otherwise, returns a list with \code{graph} and \code{summary}.
}
\description{
Constructs an undirected network based on correlation results between exposures
and features (e.g., DEGs, omics, latent factors, exposures) and optionally stores it
in the metadata of a \code{MultiAssayExperiment} object.
}
