% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_exposure.R
\name{transform_exposure}
\alias{transform_exposure}
\title{Transform Exposure Data in a MultiAssayExperiment}
\usage{
transform_exposure(
  expomicset,
  cols_of_interest = NULL,
  transform_method = "best"
)
}
\arguments{
\item{expomicset}{A \code{MultiAssayExperiment} object.}

\item{cols_of_interest}{Character vector of exposure variables to transform. If \code{NULL}, all numeric exposures are considered.}

\item{transform_method}{Transformation method to apply. Options:
\itemize{
\item \code{"none"}: No transformation.
\item \code{"log2"}: Log2 transformation.
\item \code{"x_1_3"}: Cube root transformation.
\item \code{"sqrt"}: Square root transformation.
\item \code{"best"} (default): Selects the transformation maximizing normality based on the Shapiro-Wilk test.
}}
}
\value{
A \code{MultiAssayExperiment} object with transformed exposure data.
}
\description{
Applies specified or optimal transformations to exposure data in a \code{MultiAssayExperiment}.
}
\details{
\itemize{
\item If \code{transform_method = "best"}, the function tests multiple transformations and selects the one that maximizes normality.
\item Normality is assessed using the Shapiro-Wilk test.
\item The function updates \code{colData()} with transformed values.
\item Transformation results, including normality tests, are stored in \code{metadata(expomicset)$transformation}.
}
}
\examples{
\dontrun{
expom <- transform_exposure(expom, transform_method = "best")
expom <- transform_exposure(expom, cols_of_interest = c("Exposure1", "Exposure2"), transform_method = "log2")
}

}
